+++ doc/CHANGELOG.md
@@ -26,6 +26,9 @@ commit ID.
 - Significant internal re-organisation; should not be visible to users.
 - Test: Now use -B NAME to specify images; all args after -- passed to pytest
 - Test: Now use mypy.
+- Fixed: User with UID 1000 is now removed before trying to add the dent user.
+  (Many users of dent had the same UID, and collided with images such as
+  Ubuntu that already include a user with UID 1000.)
 
 ### 1.0.0 (2024-08-29)
 - Initial release as Python distribution package.